generate_data/requirements.txt
bs4~=0.0.1
cairosvg~=2.5.2
click~=8.0.1
html2image~=2.0.1
logging~=0.4.9.6
numpy~=1.19.5
opencc~=1.1.1
pandas~=1.3.1
Pillow~=8.3.1
scikit-image~=0.18.1
